+++ src/wifi/model/yans-error-rate-model.h
@@ -22,7 +22,6 @@
 #define YANS_ERROR_RATE_MODEL_H
 
 #include "error-rate-model.h"
-#include "dsss-error-rate-model.h"
 
 namespace ns3 {
 
@@ -62,7 +61,7 @@ public:
 
   YansErrorRateModel ();
 
-  virtual double GetChunkSuccessRate (WifiMode mode, WifiTxVector txVector, double snr, uint32_t nbits) const;
+  virtual double GetChunkSuccessRate (WifiMode mode, WifiTxVector txVector, double snr, uint64_t nbits) const;
 
 
 private:
@@ -136,7 +135,7 @@ private:
    *
    * \return double
    */
-  double GetFecBpskBer (double snr, double nbits,
+  double GetFecBpskBer (double snr, uint64_t nbits,
                         uint32_t signalSpread, uint64_t phyRate,
                         uint32_t dFree, uint32_t adFree) const;
   /**
@@ -151,7 +150,7 @@ private:
    *
    * \return double
    */
-  double GetFecQamBer (double snr, uint32_t nbits,
+  double GetFecQamBer (double snr, uint64_t nbits,
                        uint32_t signalSpread,
                        uint64_t phyRate,
                        uint32_t m, uint32_t dfree,
